+++ src/components/website/team/index.js
@@ -14,3 +14,4 @@ export { default as TeamPlayerCard } from "./TeamPlayerCard";
 export { default as TeamSeasonSelector } from "./TeamSeasonSelector";
 export { default as TeamDetailTabs } from "./TeamDetailTabs";
 export { default as TeamSectionTabs } from "./TeamSectionTabs";
+export { default as TeamOverviewBackLink } from "./TeamOverviewBackLink";